WIP - create packages for OCR, Settings and Utils
