+++ app/utils/__init__.py
@@ -1,0 +1,4 @@
+from .app_logger import logger
+from .app_logger import enable_debug_logging
+
+__all__ = ["logger", "enable_debug_logging"]
